pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-trading-bot"
version = "0.1.0"
description = "Production-ready intraday crypto trading bot for Kraken"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ccxt>=4.0.0",
    "pandas>=1.5.0",
    "pandas-ta>=0.3.14b",
    "vectorbt>=0.27.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.21.0",
    "aiohttp>=3.8.0",
    "websockets>=11.0.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]